+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **Breaking**: Removed old `com.instacart.formula.RxStream`. Note: you should use `com.instacart.formula.rxjava3.RxStream`.
 - **Breaking**: Removed `Formula.start()` extension, use `Formula.toObservable()` extension instead.
 - Add `FormulaContext.onEvent` to replace `FormulaContext.callback` and `FormulaContext.eventCallback`
+- Add Kotlin coroutine support! New `formula-coroutines` module to provide `toFlow()` extension function and `FlowStream`.
 
 ## [0.7.0] - June 30, 2021
 - **Breaking**: Remove `events(observable) { }` utility function.
