+++ formula-test/src/main/java/com/instacart/formula/test/TestRuntimeExtensions.kt
@@ -2,7 +2,6 @@ package com.instacart.formula.test
 
 import com.instacart.formula.IFormula
 import com.instacart.formula.Stream
-import kotlinx.coroutines.CoroutineScope
 
 /**
  * An extension function to create a [TestFormulaObserver] for a [IFormula] instance.
@@ -14,16 +13,6 @@ fun <Input : Any, Output : Any, F: IFormula<Input, Output>> F.test(): TestFormul
 }
 
 /**
- * An extension function to create a [TestFormulaFlow] for a [IFormula] instance.
- *
- * Note: Formula won't start until you pass it an [input][TestFormulaFlow.input].
- */
-
-fun <Input : Any, Output : Any, F: IFormula<Input, Output>> F.test(scope: CoroutineScope): TestFormulaFlow<Input, Output, F> {
-    return TestFormulaFlow(scope, this)
-}
-
-/**
  * An extension function to create a [TestFormulaObserver] for a [IFormula] instance.
  *
  * @param initialInput Input passed to [IFormula].
@@ -32,20 +21,6 @@ fun <Input : Any, Output : Any, F: IFormula<Input, Output>> F.test(
     initialInput: Input
 ): TestFormulaObserver<Input, Output, F> {
     return test().apply {
-        input(initialInput)
-    }
-}
-
-/**
- * An extension function to create a [TestFormulaFlow] for a [IFormula] instance.
- *
- * @param initialInput Input passed to [IFormula].
- */
-fun <Input : Any, Output : Any, F: IFormula<Input, Output>> F.test(
-    initialInput: Input,
-    scope: CoroutineScope
-): TestFormulaFlow<Input, Output, F> {
-    return test(scope).apply {
         input(initialInput)
     }
 }
